update status after installation
