+++ components/operator/internal/state/storage_test.go
@@ -38,10 +38,16 @@ func Test_sFnStorageConfiguration(t *testing.T) {
 		next, result, err := sFnStorageConfiguration(context.Background(), r, s)
 		require.NoError(t, err)
 		require.Nil(t, result)
-		requireEqualFunc(t, sFnConfigurationStatus, next)
+		requireEqualFunc(t, sFnApplyResources, next)
 
 		require.EqualValues(t, expectedFlags, s.flagsBuilder.Build())
 		require.Equal(t, v1alpha1.StateProcessing, s.instance.Status.State)
+		requireContainsCondition(t, s.instance.Status,
+			v1alpha1.ConditionTypeConfigured,
+			metav1.ConditionTrue,
+			v1alpha1.ConditionReasonConfigured,
+			"Configuration ready",
+		)
 	})
 
 	t.Run("internal registry using azure storage", func(t *testing.T) {
@@ -95,10 +101,16 @@ func Test_sFnStorageConfiguration(t *testing.T) {
 		next, result, err := sFnStorageConfiguration(context.Background(), r, s)
 		require.NoError(t, err)
 		require.Nil(t, result)
-		requireEqualFunc(t, sFnConfigurationStatus, next)
+		requireEqualFunc(t, sFnApplyResources, next)
 
 		require.EqualValues(t, expectedFlags, s.flagsBuilder.Build())
 		require.Equal(t, v1alpha1.StateProcessing, s.instance.Status.State)
+		requireContainsCondition(t, s.instance.Status,
+			v1alpha1.ConditionTypeConfigured,
+			metav1.ConditionTrue,
+			v1alpha1.ConditionReasonConfigured,
+			"Configuration ready",
+		)
 	})
 	t.Run("internal registry using s3 storage", func(t *testing.T) {
 		s3Secret := &corev1.Secret{
@@ -161,9 +173,15 @@ func Test_sFnStorageConfiguration(t *testing.T) {
 		next, result, err := sFnStorageConfiguration(context.Background(), r, s)
 		require.NoError(t, err)
 		require.Nil(t, result)
-		requireEqualFunc(t, sFnConfigurationStatus, next)
+		requireEqualFunc(t, sFnApplyResources, next)
 
 		require.EqualValues(t, expectedFlags, s.flagsBuilder.Build())
 		require.Equal(t, v1alpha1.StateProcessing, s.instance.Status.State)
+		requireContainsCondition(t, s.instance.Status,
+			v1alpha1.ConditionTypeConfigured,
+			metav1.ConditionTrue,
+			v1alpha1.ConditionReasonConfigured,
+			"Configuration ready",
+		)
 	})
 }
