+++ components/operator/internal/state/utils.go
@@ -3,12 +3,18 @@ package state
 import (
 	"context"
 	"fmt"
+	"reflect"
+	"time"
 
 	"github.com/kyma-project/docker-registry/components/operator/api/v1alpha1"
 	"github.com/pkg/errors"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+var (
+	requeueDuration = time.Second * 3
 )
 
 func GetDockerRegistryOrServed(ctx context.Context, req ctrl.Request, c client.Client) (*v1alpha1.DockerRegistry, error) {
@@ -44,6 +50,20 @@ func GetServedDockerRegistry(ctx context.Context, c client.Client) (*v1alpha1.Do
 	}
 
 	return nil, nil
+}
+
+func updateDockerRegistryWithoutStatus(ctx context.Context, r *reconciler, s *systemState) error {
+	return r.client.Update(ctx, &s.instance)
+}
+
+func updateDockerRegistryStatus(ctx context.Context, r *reconciler, s *systemState) error {
+	if !reflect.DeepEqual(s.instance.Status, s.statusSnapshot) {
+		err := r.client.Status().Update(ctx, &s.instance)
+		emitEvent(r, s)
+		s.saveStatusSnapshot()
+		return err
+	}
+	return nil
 }
 
 func resolveRegistryHost(ctx context.Context, r *reconciler, s *systemState) (string, error) {
